recipe_api/requirements.txt
Django
djangorestframework
django-rest-knox
dj-database-url
psycopg2-binary
whitenoise
gunicorn
Pillow
django-filter
